fix(cli): split cli_smoke.rs to satisfy 300-line cap
The new `status_help_lists_v2_flags` test pushed cli_smoke.rs to 308
lines, blocking CI on the file-size guard (CONSTITUTION § 13). Move
both `status_help_lists_*` tests into a sibling `cli_smoke_status.rs`,
leaving cli_smoke.rs at 292 lines.
